Create Void Methods
Challenge Easy
Create a void method named PrintNTimes. This method should take two arguments:

A string message.
An integer n.
The method should print the message to the console n times. Use a for loop to repeat the printing.

In the Main method, call PrintNTimes with the inputs message and the input n